+++ packages/editor/src/fields/EventSelect.vue
@@ -58,8 +58,10 @@ import { TMagicButton } from '@tmagic/design';
 import type { CascaderOption, FieldProps, FormState, PanelConfig } from '@tmagic/form';
 import { MContainer, MPanel } from '@tmagic/form';
 import { ActionType } from '@tmagic/schema';
+import { DATA_SOURCE_FIELDS_CHANGE_EVENT_PREFIX } from '@tmagic/utils';
 
 import type { CodeSelectColConfig, DataSourceMethodSelectConfig, EventSelectConfig, Services } from '@editor/type';
+import { getCascaderOptionsFromFields } from '@editor/utils';
 
 defineOptions({
   name: 'MFieldsEventSelect',
@@ -99,7 +101,22 @@ const eventNameConfig = computed(() => {
         }));
       }
       if (props.config.src === 'datasource') {
-        events = dataSourceService.getEvent(formValue.type);
+        // 从数据源类型中获取到相关事件
+        events = dataSourceService.getFormEvent(formValue.type);
+        // 从数据源类型和实例中分别获取数据以追加数据变化的事件
+        const dataSource = dataSourceService.getDataSourceById(formValue.id);
+        const fields = dataSource?.fields || [];
+        if (fields.length > 0) {
+          return [
+            ...events,
+            {
+              label: '数据变化',
+              value: DATA_SOURCE_FIELDS_CHANGE_EVENT_PREFIX,
+              children: getCascaderOptionsFromFields(fields),
+            },
+          ];
+        }
+
         return events;
       }
     },
